Moving mock_notification_service_calls into top level conftest (out of integration) so we can use it for unit tests as well, and replacing the mock_send_email fixture
